requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
psycopg==3.1.8
psycopg-binary==3.1.8
pycodestyle==2.10.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.4
